Backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
black==24.10.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dill==0.3.9
distro==1.9.0
Flask==3.0.0
Flask-Cors==4.0.0
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
newsapi-python==0.2.7
newsdataapi==0.1.21
nltk==3.8.1
openai==1.12.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pylint==3.3.1
pyparsing==3.2.0
python-dotenv==1.0.1
python-magic==0.4.15
rake-nltk==1.0.6
regex==2024.11.6
requests==2.32.3
rsa==4.9
sniffio==1.3.1
tomlkit==0.13.2
tqdm==4.67.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.3
